# docs/swagger.yaml
definitions:
  apiContracts.InventoryData:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      started_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/apiContracts.UserData'
      user_id:
        type: integer
    type: object
  apiContracts.InventoryListResponse:
    properties:
      data:
        properties:
          inventories:
            items:
              $ref: '#/definitions/apiContracts.InventoryData'
            type: array
        type: object
    type: object
  apiContracts.InventoryResponse:
    properties:
      data:
        $ref: '#/definitions/apiContracts.InventoryData'
    type: object
...
  apiContracts.UserData:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.PalletEntity:
    properties:
      created_at:
        type: string
      id:
        type: integer
      inventory_id:
        type: integer
      name:
        type: string
      palletRackId:
        type: integer
      palletRackName:
        type: string
      palletizedProduct:
        items:
          $ref: '#/definitions/entities.PalletizedProductEntity'
        type: array
      qr_code:
        type: string
      qr_code_url:
        type: string
      updated_at:
        type: string
    required:
    - inventory_id
    - name
    - palletRackId
    - palletRackName
    type: object
...
  entities.PalletizedProductEntity:
    properties:
      created_at:
        type: string
      ean:
        type: integer
      id:
        type: integer
      inventory_id:
        type: integer
      palletID:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    required:
    - ean
    - quantity
    type: object
...
  inventory.InventoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  inventory.UpdateInventoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        enum:
        - Open
        - Closed
        type: string
    type: object
...
  pallet.PalletRequest:
    properties:
      name:
        type: string
      palletRackId:
        type: integer
    required:
    - name
    - palletRackId
    type: object
...
host: localhost:8000
info:
  contact:
    email: joaoferrazp@gmail.com
    name: API Support
  description: It is system to manage pallets and related products.
  title: Go Inventory API
  version: "1.0"
paths:
...
  /api/v1/inventories:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiContracts.InventoryListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all inventories
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      parameters:
      - description: Inventory Data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/inventory.InventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apiContracts.InventoryResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new inventory
      tags:
      - Inventory
  /api/v1/inventories/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiContracts.InventoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get inventory by ID
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: integer
      - description: Inventory Data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/inventory.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiContracts.InventoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an inventory
      tags:
      - Inventory
...
  /api/v1/pallets:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter by pallet rack ID
        in: query
        name: palletRackId
        type: integer
      - description: Filter by product EAN
        in: query
        name: ean
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.PalletEntity'
            type: array
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: List pallets
      tags:
      - Pallets
    post:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Pallet
      tags:
      - Pallets
  /api/v1/pallets/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Pallet
      tags:
      - Pallets
    get:
      consumes:
      - application/json
      parameters:
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get pallet by ID
      tags:
      - Pallets
    patch:
      consumes:
      - application/json
      parameters:
      - description: Palletized Product
        in: body
        name: pallet
        required: true
        schema:
          $ref: '#/definitions/pallet.PalletRequest'
      - description: ID do pallet
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PalletEntity'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Pallet
      tags:
      - Pallets
  /api/v1/pallets/export:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter by pallet rack ID
        in: query
        name: palletRackId
        type: integer
      - description: Filter by product EAN
        in: query
        name: ean
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Export pallets to CSV file
      tags:
      - Pallets
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: "Bearer {token}" para se autenticar.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"